[DEV-3843] Allow exact award id search
