+++ usaspending_api/search/tests/conftest.py
@@ -1,0 +1,4 @@
+from usaspending_api.search.tests.data.spending_by_award_test_data import spending_by_award_test_data
+
+
+__all__ = ["spending_by_award_test_data"]
